Add basic JSON parsers config
I tried to implement this with a proper language but the TOML parsing
library seems to suck, or maybe it's my fault, not sure. Anyway let's
start with the most basic thing and build up from there.

Maybe it makes more sense to configure this thing in Cue or
something?